# Restrict inode_48 growing ctor args